Replace symlink with environment variable
If users add this template to the orbit/source/extensions this will
result in symlink loops. Therefore, we replace the symlink with an
environment variable. Furthermore, we address setting a conda environment
and use ORBIT_PATH/orbit.sh -p to deal with users using conda or Isaac
Sim's bundled Python.