Replace pyproject.toml with setup.py (#10)
In order to simplify pip installations, we decided to fall back to using
`setup.py`. This PR removes `pyproject.toml` and replaces it with
`setup.py`.

An added benefit is that `setup.py` can be populated through
`config/pyproject.toml`. Therefore, a user only has to configure the
latter, rather than two files as we had previously.